% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagemeta.R
\name{addVarDescription}
\alias{addVarDescription}
\title{Add new variable}
\usage{
addVarDescription(varName, varLabel, varType, varFormat = NULL)
}
\arguments{
\item{varName}{The variable name}

\item{varLabel}{The variable label}

\item{varType}{The variable type}

\item{varFormat}{The variable format. Default is NULL}
}
\value{
None
}
\description{
Adds a new variable and corresponding characteristics to a global data frame.
}
\author{
Mango Solutions
}
