% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagemeta.R
\name{applySubsetOnLoad}
\alias{applySubsetOnLoad}
\title{Retrieves the setting as to whether or not a subset should be attached to an object on loading, 
or can actually be used to toggle the setting}
\usage{
applySubsetOnLoad(applyOnLoad)
}
\arguments{
\item{applyOnLoad}{If supplied, used to toggle the setting}
}
\value{
Setting if argument is not supplied
}
\description{
Retrieves the setting as to whether or not a subset should be attached to an object on loading, 
or can actually be used to toggle the setting
}
\author{
Mango Solutions
}

