% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamaniputils.R
\name{arrayFromMatrixList}
\alias{arrayFromMatrixList}
\title{Convert list of matrices to 3d array}
\usage{
arrayFromMatrixList(matrixList, matrixNames = names(matrixList))
}
\arguments{
\item{matrixList}{A list of matrices of the same dimension}

\item{matrixNames}{The names of the matrices, to be used as the 3rd dimension names of the array}
}
\value{
An 3d array whose submatrices come from "matrixList"
}
\description{
Creates a 3d-array from a list of matrices of the same dimension
}
\author{
Mango Solutions
}

