% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{augmentDataSubset<-}
\alias{augmentDataSubset<-}
\title{Adds subset expressions to the subset of an object}
\usage{
augmentDataSubset(x) <- value
}
\arguments{
\item{x}{Object whose subset should be modified.}

\item{value}{The subset to attach}
}
\value{
None
}
\description{
Adds subset expressions to the subset of an object
}
\author{
Mango Solutions
}
