% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagemeta.R
\name{augmentDefaultDataSubset}
\alias{augmentDefaultDataSubset}
\title{Augments the global default subset by appending the "sub" parameter to the existing subset}
\usage{
augmentDefaultDataSubset(sub)
}
\arguments{
\item{sub}{}
}
\value{
None
}
\description{
Augments the global default subset by appending the "sub" parameter to the existing subset
}
\author{
Mango Solutions
}
