% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logutils.R
\name{availableLogs}
\alias{availableLogs}
\title{Get all available logs}
\usage{
availableLogs()
}
\value{
Character vector of all available logs
}
\description{
Retrieves the names of all available logs
}
\author{
Mango Solutions
}
\keyword{programming}
\keyword{utilities,}

