% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNmLst.R
\name{cleanReportContents}
\alias{cleanReportContents}
\title{Clean report file}
\usage{
cleanReportContents(content)
}
\arguments{
\item{contents}{Lines of the report file}
}
\value{
character vector of strings without extraneous content
}
\description{
Removes extraneous content from the report file text
}
\author{
Mango Solutions
}

