% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logutils.R
\name{closeLogConnection}
\alias{closeLogConnection}
\title{Close log connection}
\usage{
closeLogConnection(logName, allowFail = TRUE)
}
\arguments{
\item{logName}{The name of the log to close}

\item{allowFail}{Logical flag.  If TRUE, will allow a closure of the log connection to fail, 
but will still disassociate the log with that connection. This is meant to be used with unclosable connections such as stdout}
}
\value{
Nothing
}
\description{
Closes a connection where a log's output goes
}
\author{
Mango Solutions
}
\keyword{IO}

