% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFortran.R
\name{convertFortran95Ops}
\alias{convertFortran95Ops}
\title{Convert FORTRAN 95 code to FORTRAN 77 code by replacing operators}
\usage{
convertFortran95Ops(code)
}
\arguments{
\item{code}{Character vector of strings of code whose operators should be converted}
}
\value{
converted codes
}
\description{
Coverts FORTRAN 95 operators in a chunk of code (in a character vector) to
}
\author{
Mango Solutions
}
