% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertNmCode.R
\name{convertNm}
\alias{convertNm}
\title{Convert NONMEM code to R}
\usage{
convertNm(model, idCol = "ID", trialCol = "TRIAL", mixCode = NULL)
}
\arguments{
\item{model}{NONMEM "code" to convert.  May be a full PRED statement}

\item{idCol}{identifier column}

\item{trialCol}{}

\item{mixCode}{}

\item{debug}{}
}
\value{
A list (or vector) of expressions holding the converted "models"
}
\description{
Converts NONMEM modelling code to executable R code, returning either a list or vector of R commands
}
\note{
Based on code originally by R. Francois
}
\author{
Mango Solutions
}

