% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{dataSubset<-}
\alias{dataSubset<-}
\title{Sets the subset of an object}
\usage{
dataSubset(x) <- value
}
\arguments{
\item{x}{Object whose subset should be modified.}

\item{value}{The subset to attach}
}
\value{
None
}
\description{
Sets the subset of an object
}
\author{
Mango Solutions
}
