% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{dataSubset}
\alias{dataSubset}
\title{Retrieves the data subset associated with a particular object}
\usage{
dataSubset(x)
}
\arguments{
\item{x}{Object (data.frame or NMProblem)}
}
\value{
Character vector of expressions
}
\description{
Retrieves the data subset associated with a particular object
}
\author{
Mango Solutions
}
