% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagemeta.R
\name{defaultDataSubset}
\alias{defaultDataSubset}
\title{Retrieves the global default subset which is assigned to objects upon loading}
\usage{
defaultDataSubset()
}
\value{
Character vector of the default subset expressions
}
\description{
Retrieves the global default subset which is assigned to objects upon loading
}
\author{
Mango Solutions
}
