% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importutils.R
\name{dirname.abs}
\alias{dirname.abs}
\title{Retrieves the full path containing some file specified in a relative
path. Simply wraps file_path_as_absolute}
\usage{
dirname.abs(x)
}
\arguments{
\item{x}{Path to the file}
}
\value{
String containing absolute path
}
\description{
Retrieves the full path containing some file specified in a relative
path. Simply wraps file_path_as_absolute
}
\author{
Mango Solutions
}

