% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagemeta.R
\name{getNmPath}
\alias{getNmPath}
\title{Get user path}
\usage{
getNmPath(pathName)
}
\arguments{
\item{pathName}{The path name}
}
\value{
The path
}
\description{
Gets a path stored globally by the user.
}
\author{
Mango Solutions
}
