% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveMisc.R
\name{getNmVersion}
\alias{getNmVersion}
\title{Gets the version of NONMEM used to create compile obj}
\usage{
getNmVersion(obj)
}
\arguments{
\item{obj}{NMProblem or NMRun}
}
\value{
A named character vector with entries "major" and "minor", corresponding to the major and minor NONMEM 
version
}
\description{
Gets the version of NONMEM used to create compile obj
}
\author{
Mango Solutions
}

