% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmegaSigma.R
\name{getSigmasOrOmegas.NM7}
\alias{getSigmasOrOmegas.NM7}
\title{Get sigmas or omegas}
\usage{
getSigmasOrOmegas.NM7(obj, what = "final", item = c("sigma", "omega"),
  method = 1, probType = c("basic", "sim"), subProblemNum = 1)
}
\arguments{
\item{obj}{NONMEM 7 problem}

\item{what}{[C,+] Objects to extract.  Equivalent to "what" for getSigmas etc.}

\item{item}{[C,1] Which item (sigma or omega) to extract}

\item{method}{[N,+] Vector of methods to extract}
}
\value{
A list or a matrix, depending on whether multiple methods and items are are selected.
}
\description{
Retrieve the omega or sigma - related items from a NONMEM 7 object.  This is an internal utility routine that was
created to refactor getSigmas and getOmegas to avoid code repetition
}
\author{
Mango Solutions
}

