% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagemeta.R
\name{getVarDescription}
\alias{getVarDescription}
\title{Get a variable}
\usage{
getVarDescription(varName)
}
\arguments{
\item{varName}{A vector of variable names}
}
\value{
A data frame with the label, format and type of each variable name
}
\description{
Gets a variable name and its associated characteristics.
}
\author{
Mango Solutions
}
