% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importutils.R
\name{hasExtension}
\alias{hasExtension}
\title{Checks if a file has a given extentions}
\usage{
hasExtension(fileName, extensions)
}
\arguments{
\item{fileName}{Name of the file to check}

\item{extension}{Extension of the file to check.  Can have more than one}
}
\value{
whether or not the filename has the given extension
}
\description{
Checks if a file has a given extentions
}
\author{
Mango Solutions
}

