% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNmLstNM7.R
\name{importMethodBlock}
\alias{importMethodBlock}
\alias{.importMethodBlock}
\title{Parse report file #METH text block}
\usage{
.importMethodBlock(methodTextBlock, SIGDIGS71 = "Sig.digs|SIG. DIGITS")
}
\arguments{
\item{methodTextBlock}{}
}
\value{
block result
}
\description{
This function parses
}
\author{
Mango Solutions
}
