% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importModelData.R
\name{importModelData}
\alias{importModelData}
\title{Import NONMEM input data}
\usage{
importModelData(dataStatement, inputStatement, dropCols = TRUE,
  trim = FALSE, path = NULL, duplicateAliased = TRUE)
}
\arguments{
\item{dataStatement}{[char matrix] - A data statement, as parsed by .importNmModData}

\item{inputStatement}{[char matrix] - A 2 column matrix describing the $INPUT statement, see .importNmModInput}

\item{dropCols}{[L,1].  If TRUE, columns  with XXXX=DROP are eliminated from returned data}

\item{trim}{Currently unused}

\item{path}{path where the data files are located}
}
\value{
A data.frame with the data inside the file
}
\description{
Imports the input data used in an individual basic NONMEM model based on its $INPUT
and $DATA statements
}
\author{
Mango Solutions
}

