% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importModelOutputTables.R
\name{importModelOutputTables}
\alias{importModelOutputTables}
\title{Import model tables}
\usage{
importModelOutputTables(tableStatement, allowFirstOnly = TRUE,
  dvLog = FALSE, trim = FALSE, returnFormat = c("singleDF", "DFlist"),
  path = NULL)
}
\arguments{
\item{tableStatement}{tableStatement [char matrix] - A control file table statement, as parsed by .importNmModTables}

\item{allowFirstOnly}{[L,1] - Allow the use of FIRSTONLY statements in the $TABLE statement?}

\item{dvLog}{currently unused}

\item{trim}{If TRUE, will _not_ add additional variables such as absWRES to the returned data}

\item{returnFormat}{How to return data. If "DFList", a list of data.frames, with one for each table file produced.
Otherwise, attempts to consolidate all of the (unique) output table variables into a single data.frame, but will return
a list of some of the tables had a FIRSTONLY statement, and others did not.}

\item{path}{Path to the table files.  Can be a path name}
}
\value{
a list or a data.frame.  
A data.frame of all unique output table columns (from all table files)
is returned if returnFormat = "singleDF", 
UNLESS there are both FIRSTONLY tables and non-FIRSTONLY tables, in which
case a list of 2 components is returned.
}
\description{
Imports the output data used in an individual NONMEM model based on its $TABLE statements
}
\author{
Mango Solutions
}
