% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNM7Iterations.R
\name{importNm7Iterations}
\alias{importNm7Iterations}
\title{Import NONMEM7 parameter iterations from .EXT files}
\usage{
importNm7Iterations(files = NULL, noTitles = NULL, noLabels = NULL,
  methods = NULL, path = ".")
}
\arguments{
\item{files}{Vectors of names of iteraton (.ext) files}

\item{noTitles}{Character vector of}

\item{noLabels}{A character vector (Should have same number of elements as \code{files}.
Should be vector of "0" and "1" entries.  
Currently, this routine will produce an error if "1" is used at all
as NOLABLES=1 is not handled yet}

\item{path}{String with path in which the files are to be found}
}
\value{
A list of data.frames of iteration information.  Each data.frame
will have been loaded via \code{importNmTable}
}
\description{
Imports a set of NONMEM 7 iterations based on NONMEM 7 control file directives.
These directives can be found next to each $EST specification (e.g. FILE=...).  
A list of data.frames of iterations will be returned. 
This routine is not meant to be used directly by an end user
}
\author{
Mango Solutions
}
