% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNmInter.R
\name{importNmInter}
\alias{importNmInter}
\title{Import NONMEM "INTER" file}
\usage{
importNmInter(fileName, path = NULL)
}
\arguments{
\item{fileName}{[C,1] Name of the "Inter" file to read}

\item{path}{[C,1] Path that holds the file}
}
\value{
A data.frame with the iteration information
}
\description{
Imports a NONMEM "INTER" file into a data.frame with one row for each iteration
}
\note{
Based on code by R. Pugh,J.James, and R.Francois
}
\author{
Mango Solutions
}
