% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNmLstMatrix.R
\name{importNmLstMatrix}
\alias{importNmLstMatrix}
\alias{.importNmLstMatrix}
\title{Extract matrix from a report file}
\usage{
.importNmLstMatrix(txt, condensed = NULL, onlyDiag = FALSE)
}
\arguments{
\item{txt}{[C,+] character vector of lines of text}

\item{condensed}{[L,1] Logical flag.  If TRUE, will use "condensed" matrix notation}

\item{onlyDiag}{[L,1] Logical flag. If TRUE, will only obtain the diagonal of the matrix}
}
\value{
A matrix with the contents of the matrix described by txt
}
\description{
Obtains a matrix from a chunk of report file text (such as the estimates of sigma or omega).  
It is meant to be used on elements of the list returned by sectionLst
}
\note{
Based on code by R. Francois
}
\author{
Mango Solutions
}
