% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNmLst.R
\name{importNmLstSimDataGen}
\alias{importNmLstSimDataGen}
\title{Import report file contents of a simulation data generation problem ( without model fitting step )}
\usage{
importNmLstSimDataGen(contents)
}
\arguments{
\item{contents}{[C,+] Character vector of}
}
\value{
A list with 2 elements: nRecords and nIndividuals, holding the number of individuals and number of
records in the problem respectively
}
\description{
Imports the contents of a simulation data generation problem from a report file into a list
}
\author{
Mango Solutions
}

