% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNmLstNM7.R
\name{importNmLstSimModel.NM7}
\alias{importNmLstSimModel.NM7}
\title{import NONMEM7 simulation+model fit problem report contents}
\usage{
importNmLstSimModel.NM7(contents, numSub = NA)
}
\arguments{
\item{[C, +]}{contents Contents of the report file for a single problem}

\item{[N, }{1] numSub Number of sub-problems in the problem}
}
\value{
out list
}
\description{
Imports the contents of the report file of a SimModel type problem compiled with NONMEM 7
}
\author{
Mango Solutions
}

