% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNmLst.R
\name{importNmLstSimModel}
\alias{importNmLstSimModel}
\title{Import simulation+fit report}
\usage{
importNmLstSimModel(contents, numSub = NA)
}
\arguments{
\item{contents}{Contents of the report file for a single problem}

\item{numSub}{Number of sub-problems in the problem}
}
\value{
A list with parsed elements
}
\description{
Imports the contents of the report file of a SimModel type problem
}
\author{
Mango Solutions
}

