% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNmLstVector.R
\name{importNmLstVector}
\alias{importNmLstVector}
\alias{.importNmLstVector}
\title{Reads a "vector" stored in the text of a list file}
\usage{
.importNmLstVector(text)
}
\arguments{
\item{text}{Vector of strings of text of a list file}
}
\value{
A named numeric vector
}
\description{
Reads a "vector" stored in the text of a list file
}
\author{
Mango Solutions
}
