% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNmModInput.R
\name{importNmModInput}
\alias{importNmModInput}
\alias{.importNmModInput}
\title{Parse $INPUT statement}
\usage{
.importNmModInput(txt, .extract = length(grep("\\\\$INPUT", toupper(txt))) >
  0)
}
\arguments{
\item{txt}{Character vector of text containing an $INPUT statement, typically the 
contents of a control file}

\item{.extract}{Flag indicating whether this is the $INPUT section itself, or whether the section must first be
extracted}
}
\value{
a 2 column matrix mapping the variables in the input
statement with the variables used by NONMEM, which can be different by setting
aliases, e.g. ID=SUBJ in the $INPUT statement.  Also some may be dropped
}
\description{
Parses the $INPUT section of a NONMEM control file
}
\examples{
.importNmModInput("$INPUT  ID DOSE=AMT TIME CP=DV")
}
\author{
Mango Solutions
}
