% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNmModSingleProblem.R,
%   R/importNmModSingleProblemNM7.R
\name{importNmModSingleProblem}
\alias{importNmModSingleProblem}
\alias{.importNmModSingleProblem}
\alias{importNmModSingleProblem}
\alias{.importNmModSingleProblemNM7}
\title{Import Single NONMEM problem}
\usage{
.importNmModSingleProblem(contents, fileName)

.importNmModSingleProblemNM7(contents, fileName)
}
\arguments{
\item{contents}{Text of the contents of a SINGLE problem statement}

\item{fileName}{Name of the control file from which the problem originates}

\item{contents}{Text of the contents of a SINGLE problem statement}

\item{fileName}{Name of the control file from which the problem originates}
}
\value{
named list with the various sections split up

named list with the various sections split up
}
\description{
Imports a single NONMEM problem into a list structure

Imports a single NONMEM problem into a list structure
}
\author{
Mango Solutions

Mango Solutions
}
