% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNmModTheta.R
\name{importNmModTheta}
\alias{importNmModTheta}
\alias{.importNmModTheta}
\title{Parse $THETA statement}
\usage{
.importNmModTheta(txt = NULL, guessNames = TRUE, rx = "([^~[:space:]]+)$",
  fileName = NULL)
}
\arguments{
\item{txt}{[C,+] - Vector of strings in which to extract THETAs}

\item{guessNames}{[L,1] - Try to deduce the names of the thetas (CL, KV, etc.)}

\item{rx}{Regular expression used for detecting names of thetas}

\item{fileName}{}
}
\value{
A matrix with one row for each "theta" and 3 columns : 1 for the lower value bounds.
one for the initial estimate, and one for the upper bound
}
\description{
Extracts the "theta" declaration information from the text of a control file
}
\author{
Mango Solutions
}
