% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importNmLstNM7.R
\name{importNmReport.NM7}
\alias{importNmReport.NM7}
\title{Import NONMEM 7 report file}
\usage{
importNmReport.NM7(content, textReport = FALSE)
}
\arguments{
\item{content}{[C,+] The report file's text.}

\item{textReport}{[L,1] Should standard text messages be logged to the stdReport log?}
}
\value{
A list with various parsed components of the report file.
}
\description{
Parses the contents of a NONMEM 7 report file into a list of parsed components 
(see the design info for details on the return structure)
}
\author{
Mango Solutions
}

