% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logutils.R
\name{logConnection}
\alias{logConnection}
\title{Get log's connection}
\usage{
logConnection(logName)
}
\arguments{
\item{logName}{The name of the log}
}
\value{
Connection associated to a log
}
\description{
Retrieves the connection associated with a particular log
}
\author{
Mango Solutions
}
\keyword{utilities}

