% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logutils.R
\name{logMessage}
\alias{logMessage}
\title{Message logging}
\usage{
logMessage(text, logName)
}
\arguments{
\item{text}{Character vector to output}

\item{logName}{Name of the log.  Should be one of the available logs}
}
\value{
None
}
\description{
Writes a character vector to some given log, if that log is open
}
\author{
Mango Solutions
}
\keyword{IO,}
\keyword{debugging,}
\keyword{utilities}

