% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseutils.R
\name{lseq}
\alias{lseq}
\title{generates a set of sequences that start at x[i] and end at x[i+1], where i varies from 1 to length(x)-1.  If txt is not
NULL, in addition it will subsets of txt created by indexing it with these sequences}
\usage{
lseq(x, txt = NULL)
}
\arguments{
\item{x}{integer vector}

\item{txt}{}
}
\value{
List of sequences, as described above
}
\description{
generates a set of sequences that start at x[i] and end at x[i+1], where i varies from 1 to length(x)-1.  If txt is not
NULL, in addition it will subsets of txt created by indexing it with these sequences
}
\author{
Mango Solutions
}

