% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMBasicModelNM7_factory.R
\name{NMBasicModelNM7}
\alias{NMBasicModelNM7}
\title{NMBasicModel constructor}
\usage{
NMBasicModelNM7(controlStatements, path, reportContents,
  dropInputColumns = FALSE, versionInfo = c(major = "VII", minor = 1),
  conFile = NULL)
}
\arguments{
\item{controlStatements}{[list] A list of control file statements for this particular problem}

\item{path}{[C,1] path parameter, passed directly to importModelData and importModelOutputTables}

\item{versionInfo}{[C, +] numeric vector that holds}

\item{lstContents}{[list] contents of an lst file that apply to this problem}
}
\value{
An NMBasicModel object holding the problem information
}
\description{
Constructs an NMBasicModel object from the control statements and output list statements that apply to it.
Meant to be used from within importNm
}
\author{
Mango Solutions
}
