% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_NMClasses.R
\docType{class}
\name{NMProblem-class}
\alias{NMProblem-class}
\title{An S4 class of NMProblem}
\description{
An S4 class of NMProblem
}
\section{Slots}{

\describe{
\item{\code{problemStatement}}{The contents of the $PROB statement}

\item{\code{controlStatements}}{A list of parsed sections of the control file}

\item{\code{reportStatements}}{A list of parsed sections of the output report file}

\item{\code{nmVersionMajor}}{[C,1] NONMEM major version number}

\item{\code{nmVersionMinor}}{[N,1] NONMEM minor version number}

\item{\code{inputData}}{A data.frame of the input data, if available (otherwise an empty data.frame)}

\item{\code{outputData}}{Aggregation of the output data}

\item{\code{additionalVars}}{A data.frame of additional variables created by the user}
}}

