% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_NMClasses.R
\docType{class}
\name{NMRun-class}
\alias{NMRun-class}
\title{This is the basic class for handling a standard NONMEM run.  In essence, it holds
several individual problems which may be of type NMBasicModel, NMSimModel, and NMSimDataGen.
It also holds the control file and list output file text}
\description{
This is the basic class for handling a standard NONMEM run.  In essence, it holds
several individual problems which may be of type NMBasicModel, NMSimModel, and NMSimDataGen.
It also holds the control file and list output file text
}
\section{Slots}{

\describe{
\item{\code{controlText}}{[C,+] Text of the control file, without comments}

\item{\code{controlComments}}{[C,1] Comments of each line of the control file}

\item{\code{nmVersionMajor}}{[C,1] NONMEM major version number}

\item{\code{nmVersionMinor}}{[N,1] NONMEM minor version number}

\item{\code{controlFileInfo}}{[data.frame] Information about the control file}

\item{\code{reportFileInfo}}{[data.frame] Information about the list file}

\item{\code{numProblems}}{[N,1] - Number of problems in the run}

\item{\code{problems}}{[list] - List of the actual problem results}

\item{\code{reportText}}{Text of the lst output file}
}}

