\name{nmRunReport-class}
\docType{class}
\alias{nmRunReport-class}

\title{Auxilliary classes for representing report file statements}
\description{Auxilliary classes for representing report file statements}

\section{Objects from the Class}{
Objects of this class inherit from the \code{\link{list}} class .
}

\section{Extends}{
Class \code{\link{list}}, directly.
}

\author{Mango Solutions <support@mango-solutions.com>}

\examples{
	 showClass("nmRunReport")
}
\keyword{classes}
