% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_NMClasses.R
\docType{class}
\name{NMSimDataGen-class}
\alias{NMSimDataGen-class}
\title{This class handles NONMEM models that have a $SIM statement but only simulate
data without fitting a model}
\description{
This class handles NONMEM models that have a $SIM statement but only simulate
data without fitting a model
}
\section{Slots}{

\describe{
\item{\code{numSimulations}}{Number of simulations/subproblems}

\item{\code{seeds}}{Seeds used for the random number generator}

\item{\code{problemStatement}}{$PROBLEM statement}

\item{\code{thetaInitial}}{Fixed/initial theta values used to generate data}

\item{\code{omegaInitial}}{Fixed/initial omega values used to generate data}

\item{\code{sigmaInitial}}{Fixed/initial sigma values used to generate data}
}}

