% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMSimModel_factory.R
\name{NMSimModel}
\alias{NMSimModel}
\title{Constructor for the NMSimModel class}
\usage{
NMSimModel(controlStatements, path, reportContents, versionInfo = c(major =
  "VI", minor = 0))
}
\arguments{
\item{controlStatements}{Set of parsed control statements}

\item{path}{Path where the run is located}

\item{reportContents}{Parsed contents of the report file}
}
\value{
Newly constructed object
}
\description{
Constructor for the NMSimModel class
}
\author{
Mango Solutions
}

