% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_NMClasses.R
\docType{class}
\name{NMSimModelNM7-class}
\alias{NMSimModelNM7-class}
\title{This class holds the results of NONMEM problems that both simulate data and fit the model during each simulation
(that is, models with a $SIM statement in the control file but no "ONLY" keyword).}
\description{
This class holds the results of NONMEM problems that both simulate data and fit the model during each simulation
(that is, models with a $SIM statement in the control file but no "ONLY" keyword).
}
\section{Slots}{

\describe{
\item{\code{numSimulations}}{Number of simulations generated}

\item{\code{objectiveFinal}}{Vector of final values of the objective functions}

\item{\code{thetaFinal}}{Final estimates of the "thetas", stored as a matrix with one row for each simulation}

\item{\code{omegaFinal}}{Final estimates of the "omegas", stored as a 3-d array with one matrix for each simulation}

\item{\code{sigmaFinal}}{Final estimates of the "sigmas", stored as a 3-d array with one matrix for each simulation}

\item{\code{thetaInitial}}{Initial values of thetas}

\item{\code{omegaInitial}}{Initial values of omegas}

\item{\code{sigmaInitial}}{Initial values of sigmas}

\item{\code{seeds}}{Values of seeds used for random-number generation}

\item{\code{numMethods}}{TODO}

\item{\code{methodInfo}}{TODO}

\item{\code{methodNames}}{TODO}
}}

