% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMSimModelNM7_factory.R
\name{NMSimModelNM7}
\alias{NMSimModelNM7}
\title{create NONMEM 7 simulation+fitting class}
\usage{
NMSimModelNM7(controlStatements, path, reportContents, versionInfo = c(major =
  "VII", minor = 1))
}
\arguments{
\item{controlStatements}{Set of parsed control statements}

\item{path}{Path where the run is located}

\item{reportContents}{Parsed contents of the report file}
}
\value{
Newly constructed object of class NMSimModelNM7
}
\description{
Constructor for the NMSimModelNM7 class
}
\author{
Mango Solutions
}

