% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstparseutiles.R
\name{nmTitle}
\alias{nmTitle}
\title{split list text into sections}
\usage{
nmTitle(txt, titleSecLength = 5, maxTitleLines = 2)
}
\arguments{
\item{txt}{[C,+] The input sectioned chunk of the lst file.  This is typically obtained from sectionLst or
sectionMethodBlocks.  See the comment block below for examples of section text}

\item{titleSecLength}{[N,1]}

\item{maxTitleLines}{[N,1]}
}
\value{
a list with two elements: "text" and "title".  The latter is the title of the section 
(e.g. "FINAL PARAMETER ESTIMATE"), the former is the text content without the titles
}
\description{
Splits a chunk of report file section into its title/header and its title.
}
\author{
Mango Solutions
}

