% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstparseutiles.R
\name{nmVersion}
\alias{nmVersion}
\title{Obtains the NONMEM version information from a list file}
\usage{
nmVersion(lstContents)
}
\arguments{
\item{lstContents}{[C,+] Contents of the lst file}
}
\value{
A list with two elements: the version and the level
}
\description{
Obtains the NONMEM version information from a list file
}
\author{
Mango Solutions
}

