% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.table.statements.R
\name{normalize.table.statements}
\alias{normalize.table.statements}
\title{For a table statement, if user put multiple un-mergeble tables into one text output file
We split it into multiple table files as [origianlname]__1, [origianlname]__2 , e.t.c
to allow the importing function working correctly}
\usage{
normalize.table.statements(tableStatement, force.change = FALSE)
}
\arguments{
\item{tableStatement}{a tableStatement in the control stream, typically returned by importNmMod}

\item{force.change}{return the normalized tableStatement even the original file is not splited successfully}
}
\value{
the normalized tableStatement which each table resides in its own file
}
\description{
For a table statement, if user put multiple un-mergeble tables into one text output file
We split it into multiple table files as [origianlname]__1, [origianlname]__2 , e.t.c
to allow the importing function working correctly
}

