% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseutils.R
\name{ogrep}
\alias{ogrep}
\title{"grep -o"}
\usage{
ogrep(pattern, text, filter = NULL, ignore.case = TRUE, perl = FALSE)
}
\arguments{
\item{pattern}{regular expression to search for}

\item{text}{vector of strings in which to search}

\item{filter}{[C,1] replaces matched instances with this string.}

\item{ignore.case}{makes regexp match case-insensitive if true}

\item{perl}{logical flag - use perl conventions for string matching?}
}
\value{
A vector of segments of texts that match "pattern"
}
\description{
similiar to "grep -o" in UNIX: returns only the sections of a vector of 
strings that match a regular expression
}
\author{
Mango Solutions
}

