% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstparseutiles.R
\name{partitionLstSubproblems}
\alias{partitionLstSubproblems}
\title{partitionLstSubproblems}
\usage{
partitionLstSubproblems(contents)
}
\arguments{
\item{contents}{A character vector of strings holding the lines of subproblem information (ONLY!) for one simulation problem}
}
\value{
list
}
\author{
Mango Solutions
}

