% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstparseutiles.R
\name{partitionMethods}
\alias{partitionMethods}
\title{Partitions the text of a report file into chunks of text for each "method" used.  This applies to
NONMEM 7 reports only, as they will now have separate parameter estimates, objective function values etc.
for different $EST methods employed.}
\usage{
partitionMethods(lstProblemContents)
}
\arguments{
\item{lstProblemContents}{A character vector of lines of text from a NONMEM 7 report file}
}
\value{
A list of character vectors of split text.  Each element of the list will also have an attribute
called "method.name".  This will denote the actual name of the method used for that chunk.
}
\description{
Partitions the text of a report file into chunks of text for each "method" used.  This applies to
NONMEM 7 reports only, as they will now have separate parameter estimates, objective function values etc.
for different $EST methods employed.
}
\author{
Mango Solutions
}

