% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importutils.R
\name{processPath}
\alias{processPath}
\title{"Processes" a path, so that if it is enclosed in brackets, it is taken to be the name of a path
to be retrieved by "getNmPath".  Otherwise, just returns the path}
\usage{
processPath(path)
}
\arguments{
\item{path}{the path to process.}
}
\value{
Either path unmodified, or path retrieved via getNmPath
}
\description{
"Processes" a path, so that if it is enclosed in brackets, it is taken to be the name of a path
to be retrieved by "getNmPath".  Otherwise, just returns the path
}
\author{
Mango Solutions
}

