% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagemeta.R
\name{removeNmPath}
\alias{removeNmPath}
\title{Remove user path}
\usage{
removeNmPath(pathName)
}
\arguments{
\item{pathName}{The path name}
}
\value{
None
}
\description{
Removes a globally stored user path.
}
\author{
Mango Solutions
}
