% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorhandling.R
\name{RNMImportStop}
\alias{RNMImportStop}
\title{Generate an exception and message}
\usage{
RNMImportStop(msg = "Error!\\n", call = NULL)
}
\arguments{
\item{msg}{Message}

\item{call}{(optional).  Call information for function that generated exception (i.e. from match.call)}
}
\value{
None
}
\description{
Generates a formal exception, and prints the sequence of calls that generated this expression.  This function
is meant to provide custom error handling for this package.
}
\author{
Mango Solutions
}
\keyword{error}

