% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorhandling.R
\name{RNMImportWarning}
\alias{RNMImportWarning}
\title{Generate a warning}
\usage{
RNMImportWarning(msg = "Warning!\\n", call = NULL)
}
\arguments{
\item{msg}{}

\item{call}{}
}
\value{
None
}
\description{
Generates a formal warning. This function is meant to provide custom warning generation for this package.
}
\author{
Mango Solutions
}
\keyword{error}

