% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importutils.R
\name{scanFile}
\alias{scanFile}
\title{Read file contents and strip whitespace}
\usage{
scanFile(file, empty.rx = "^[[:space:]]*$")
}
\arguments{
\item{file}{name of the file to read}

\item{empty.rx}{regular expression to eliminate from strings read in}
}
\value{
Character vectors of lines of file contents
}
\description{
Reads the lines of a file and removes whitespace in the process
}
\author{
Mango Solutions
}
\keyword{IO}

