% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{section}
\alias{section}
\title{section}
\usage{
section(text, startSection = NULL, endSection = NULL, oneline = FALSE,
  startpattern = sprintf("^\%s(\%s)", sectionSep, paste(startSection, collapse
  = "|")), endpattern = sprintf("^\%s(\%s)", sectionSep, paste(endSection,
  collapse = "|")), stripout = oneline, as.list = TRUE,
  remove.comments = FALSE, glue = FALSE, sectionSep = "\\\\$",
  clean = FALSE, ...)
}
\arguments{
\item{text}{control file contents}

\item{startSection}{[C,1] token that demarcates beginning of the section to process necessary because some sections 
have multiple entries (e.g. multiple $THETA sections)}

\item{endSection}{[C,1] token for section end}

\item{oneline}{some sections should end up occupying only one line.  this is a flag that forces this}

\item{startpattern}{[C,1] full pattern for section start regexp searches, generally unused as "startSection" is sufficient}

\item{stripout}{logical flag - remove the $* bits ?}

\item{as.list}{logical flag - return section result as a list?}

\item{remove.comments}{- self explanatory}

\item{glue}{}

\item{sectionSep}{[C, 1] token that precedes section breaks}

\item{clean}{}

\item{...}{}

\item{endPattern}{[C,1] full pattern for section end regexp searches, generally unused as "startSection" is sufficient}
}
\value{
A list of character vectors, where each element of the list is a section of the control file
}
\description{
Used to split control file into sections.  Sections are determined
by checking for "$"
}
\author{
Mango Solutions
}

