% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstparseutiles.R
\name{sectionLst}
\alias{sectionLst}
\title{Partitions the contents of a list file into sections that can then be parsed individually}
\usage{
sectionLst(fileContents, sep, keep.all = FALSE, recurse = TRUE)
}
\arguments{
\item{fileContents}{[C,+] Lines of text from a list file}

\item{sep}{(?) passed to .cleanRx}

\item{keep.all}{(?)}

\item{recurse}{Should the function be called recursively?}
}
\value{
A (partially) named list with relevant sections of the lst file.
}
\description{
Partitions the contents of a list file into sections that can then be parsed individually
}
\note{
Based on code by R. Francois
}
\author{
Mango Solutions
}

