% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstparseutiles.R
\name{sectionMethodBlock}
\alias{sectionMethodBlock}
\title{Split method text block into sections}
\usage{
sectionMethodBlock(methodContents)
}
\arguments{
\item{[C, +]}{methodContents Line of text in a #METH statement (includes the #METH line)}
}
\value{
[L, +] a named list of split text, with names corresponding to section names
}
\description{
Splits a single block of #METH statements into "sections".  This is similar to sectionLst, but only for a
single block delimeted by #METH in NONMEM 7
}
\author{
Mango Solutions
}

