% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagemeta.R
\name{setDefaultDataSubset}
\alias{setDefaultDataSubset}
\title{Modifies the global default subset for loaded objects, and can toggle if it is applied on loading}
\usage{
setDefaultDataSubset(sub, applyOnLoad)
}
\arguments{
\item{sub}{character vector of strings of subset expressions}

\item{applyOnLoad}{Logical flag.  Should the subset actually be applied on loading automatically?}
}
\value{
None
}
\description{
Modifies the global default subset for loaded objects, and can toggle if it is applied on loading
}
\author{
Mango Solutions
}
