% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logutils.R
\name{setLogConnection}
\alias{setLogConnection}
\title{Set log connection}
\usage{
setLogConnection(logName, conn)
}
\arguments{
\item{logName}{Name of the log to modify}

\item{conn}{Connection}
}
\value{
Nothing
}
\description{
Sets a log's output connection to a given connection
}
\author{
Mango Solutions
}
\keyword{IO,}
\keyword{debugging,}
\keyword{utilities}

