% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logutils.R
\name{setLogFile}
\alias{setLogFile}
\title{Set log connection to a file}
\usage{
setLogFile(logName, fileName)
}
\arguments{
\item{logName}{Name of the log}

\item{fileName}{Name of the file (need not exist)}
}
\value{
None
}
\description{
Points a given log to a file connection
}
\author{
Mango Solutions
}
\keyword{IO}

