% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagemeta.R
\name{setNmPath}
\alias{setNmPath}
\title{Set new path}
\usage{
setNmPath(pathName, path)
}
\arguments{
\item{pathName}{The path name}

\item{path}{The path}
}
\value{
None
}
\description{
Sets a new path globally to be used by the user.
}
\author{
Mango Solutions
}
