% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagemeta.R
\name{setVarDescription}
\alias{setVarDescription}
\title{Set new variable}
\usage{
setVarDescription(varName, varLabel = NULL, varFormat = NULL,
  varType = NULL)
}
\arguments{
\item{varName}{The variable name}

\item{varLabel}{A new label for the variable. Default is NULL}

\item{varFormat}{A new format for the variable. Default is NULL}

\item{varType}{A new type for the variable. Default is NULL}
}
\value{
None
}
\description{
Sets new characteristics for a required variable.
}
\author{
Mango Solutions
}
