% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_methods.R
\name{showOutput}
\alias{showOutput}
\title{Show output table info}
\usage{
showOutput(object)
}
\arguments{
\item{object}{NMProblem object}
}
\value{
none
}
\description{
Displays information about output data.  Utility function for use with show methods
}
\author{
Mango Solutions
}

