% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.table.statements.R
\name{split.file.into}
\alias{split.file.into}
\title{split a multiple table text file into multiple files which contain only one for each}
\usage{
\method{split}{file.into}(fn, num)
}
\arguments{
\item{fn}{the filename of multiple table text file}

\item{num}{how many tables need to be splitted into}
}
\value{
logical value indicating if splitting is successfully
}
\description{
split a multiple table text file into multiple files which contain only one for each
}

