% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseutils.R
\name{splitVector}
\alias{splitVector}
\title{split a vector into a list of sub-vectors}
\usage{
splitVector(x, indices, includeStart = FALSE, includeEnd = FALSE)
}
\arguments{
\item{x}{The vector to split}

\item{indices}{a set of integer indices along which to split}

\item{includeStart}{should the first element of the vector be included automatically?}

\item{includeEnd}{should the last element of the vector be included automatically?}
}
\value{
A list of split elements of x.  These will be of the form x[index_n:(index_n - 1)] for each index
}
\description{
"Splits" a given vector along a vector of indices. The split content is returned as a list ranging between each index,
with each index
}
\author{
Mango Solutions
}

