% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseutils.R
\name{stripBlanks}
\alias{stripBlanks}
\title{stripBlanks}
\usage{
stripBlanks(txt, trim = TRUE, leading = TRUE, trailing = TRUE,
  shrink = TRUE, trimRx = "[[:space:]]+", trailingRx = "[[:space:]]*$",
  leadingRx = "^[[:space:]]*", remove.empty = FALSE)
}
\arguments{
\item{txt}{[C,+] the text to modify}

\item{trim}{[L,1] replace multi space with single space? (default TRUE)}

\item{leading}{[L,1] remove leading spaces? (default TRUE)}

\item{trailing}{[L,1] remove trailing spaces? (default TRUE)}

\item{shrink}{[L,1] shrink spaces around '=' signs? (default TRUE)}

\item{trimRx}{regular expression used if trim}

\item{trailingRx}{regular expression used if trailing}

\item{leadingRx}{regular expression used if leading}

\item{remove.empty}{[L,1] remove empty elements? (default FALSE)}
}
\value{
Character vector
}
\description{
Removes whitespace from character vector.
Multiple whitespace characters are consolidated to a single space.
}
\examples{
stripBlanks(character(0))
stripBlanks(c("", "  "))
stripBlanks("  AAA")
stripBlanks("  AAA   ")
stripBlanks(c("  AAA   ", "   BBB   CCC"))
stripBlanks(c("\\tAAA   ", "\\tBBB\\nCCC"))
stripBlanks(c("  AAA   ", "   "), remove.empty = TRUE)
}
\author{
Mango Solutions
}
