% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addedData.R
\docType{methods}
\name{addedData}
\alias{addedData}
\alias{addedData,NMRun-method}
\alias{addedData,NMProblem-method}
\title{Get added data}
\usage{
addedData(obj, ...)

\S4method{addedData}{NMRun}(obj, problemNum = 1)

\S4method{addedData}{NMProblem}(obj)
}
\arguments{
\item{obj}{An object of class NMRun or extending NMProblem}

\item{...}{problemNum for addedData.NMRun}
}
\value{
A data.frame containing the derived variables
}
\description{
Obtain the derived/user-added variables from a NONMEM problem object
}
\details{
Most of these added variables will probably be created via
addDerivedCategorical
}
\examples{
x <- importNm("theoph.con", path = system.file("examples/theoph", package = "RNMImport") ) 
	y <- getProblem(x)
	y <- addDerivedCategorical(y, "DV", breaks = 3, labels = c("low", "medium", "high"))
	y <- addDerivedCategorical(y, "TIME", breaks = 6)
	print(addedData(y))
}
\author{
Mango Solutions
}
\keyword{utils}
