\name{addVarDescription}
\alias{addVarDescription}
\title{Adds a variable description}
\description{Adds a new variable and appropriate characteristics to a global data frame for future reference, e.g. one may want to
add a variable DOSAGEMG.}
\usage{
addVarDescription(varName, varLabel, varType, varFormat = NULL)
}
\arguments{
  \item{varName}{The variable name}
  \item{varLabel}{The variable label}
  \item{varType}{The variable type}
  \item{varFormat}{The variable format. NULL by default}
}

\value{None}
\author{Mango Solutions <rweeks@mango-solutions.com>}
\keyword{environment}