% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{applyDataSubset}
\alias{applyDataSubset}
\title{Applies a set of subset expressions to a data.frame}
\usage{
applyDataSubset(obj, sub = NULL, verboseWarnings = FALSE)
}
\arguments{
\item{obj}{A data.frame to be subsetted}

\item{sub}{Character vector of subsets to apply}

\item{verboseWarnings}{[L, 1] Logical flag.  If TRUE, will emit a warning whenever a subset statement fails}
}
\value{
subsetted data.frame, with subsets in sub applied
}
\description{
Applies a set of subset expressions to a data.frame
}
\author{
Mango Solutions
}
