% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveMisc.R
\docType{methods}
\name{getControlStatements}
\alias{getControlStatements}
\alias{getControlStatements,NMRun-method}
\alias{getControlStatements,NMProblem-method}
\title{Extract list of parsed control statement}
\usage{
getControlStatements(obj, ...)

\S4method{getControlStatements}{NMRun}(obj, problemNum = 1)

\S4method{getControlStatements}{NMProblem}(obj, ...)
}
\arguments{
\item{obj}{NMRun or NMProblem object}

\item{...}{Various arguments specific to methods}
}
\value{
A list with the parsed control statements (e.g. an actual matrix for the Thetas, etc.,
}
\description{
Extracts the list of "parsed control statements" associated with a problem or run
}
\author{
Mango Solutions
}
