% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveMisc.R
\name{getFileinfo}
\alias{getFileinfo}
\alias{getReporttext}
\alias{getControltext}
\title{Retrieve file information}
\usage{
getFileinfo(run)

getReporttext(run)

getControltext(run)
}
\arguments{
\item{run}{Object of class NMRun}
}
\value{
A data.frame with 2 rows, 1 describing the report file and the other the control file
}
\description{
Retrieves information about a run's control and report files as a data.frame
}
\author{
Mango Solutions
}
\keyword{utility}
