% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethodNames.R
\docType{methods}
\name{getMethodNames}
\alias{getMethodNames}
\alias{getMethodNames,NMRun-method}
\alias{getMethodNames,NMBasicModelNM7-method}
\alias{getMethodNames,NMSimModelNM7-method}
\title{Get method names}
\usage{
getMethodNames(obj, what = c("report", "control"), problemNum = 1)

\S4method{getMethodNames}{NMRun}(obj, what = c("report", "control"),
  problemNum = 1)

\S4method{getMethodNames}{NMBasicModelNM7}(obj, what = c("report", "control"),
  problemNum = 1)

\S4method{getMethodNames}{NMSimModelNM7}(obj, what = c("report", "control"),
  problemNum = 1)
}
\arguments{
\item{obj}{NMRun, NMBasicModelNM7, or NMSimModelNM7}

\item{what}{The string "report" or "control" - selects which file should be used
for the method names}

\item{problemNum}{Problem number in run}
}
\value{
A character vector of method names used
}
\description{
Extracts estimation method names from a run or problem that used NONMEM 7.  These
can be the names used in the control file, or the report file
}
\note{
If what = "control", the names are taken from the $EST control sections
}
\examples{
\dontrun{
     x <- importNm("TestData1.ctl", "TestData1.lst", path  = "testing/testdata/TestDataNM7")
     print(getMethodNames(x))
}
}
\author{
Mango Solutions
}
\keyword{methods}
\keyword{utilities}
