% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfileextensions.R
\name{getNmFileExtensions}
\alias{getNmFileExtensions}
\title{Gets allowed NONMEM file extensions}
\usage{
getNmFileExtensions(parameter)
}
\arguments{
\item{parameter}{String with the file type. Must be one of "control", "report", "table" or "input"}
}
\value{
A vector of allowed file type extensions
}
\description{
Returns a vector of allowed file extensions for a specific NONMEM file type, e.g. control, report, output table or 
input data.Gets the conventional file name extensions used by NONMEM.
}
\examples{
getNmFileExtensions("control")
}
\author{
Mango Solutions
}
\keyword{environment}
