\name{getNmPath}
\alias{getNmPath}
\title{Gets a user defined path}
\description{Gets a path which is stored globally as a user specified path name.}
\usage{
getNmPath(pathName)
}
\arguments{
  \item{pathName}{The name of the path}
}
\value{The path corresponding to the path name}
\author{Mango Solutions <rweeks@mango-solutions.com>}
\keyword{environment}
