% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProblem.R
\name{getProblem}
\alias{getProblem}
\title{Extract an individual NONMEM problem}
\usage{
getProblem(run, problemNumber = 1)
}
\arguments{
\item{run}{An NMRun class object}

\item{problemNumber}{Number of the problem to retrieve}
}
\value{
The given problem number (NMBasicModel, NMSimDataGen, or NMSimModel)
}
\description{
Retrieves a desired problem from a full NONMEM run object.
}
\note{
If \code{problemNum} is out of the acceptable range of problems, an exception will be generated.  The same
applies if \code{run} is not an NMRun object
}
\author{
Mango Solutions
}
\keyword{manip}
