% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSigmas.R
\docType{methods}
\name{getSigmas}
\alias{getSigmas}
\alias{getSigmas,NMBasicModel-method}
\alias{getSigmas,NMRun-method}
\alias{getSigmas,NMSimModel-method}
\alias{getSigmas,NMSimModelNM7-method}
\alias{getSigmas,NMSimDataGen-method}
\title{A generic function that extracts sigma estimates (and initial estimates and standard errors if specified) from a NONMEM object.
Shrinkage values may also be retrieved for NONMEM 7 objects.}
\usage{
getSigmas(obj, what = "final", subProblemNum = 1, method = 1,
  problemNum = 1)

\S4method{getSigmas}{NMBasicModel}(obj, what = "final", subProblemNum = 1,
  method = 1, problemNum = 1)

\S4method{getSigmas}{NMRun}(obj, what = "final", subProblemNum = 1,
  method = 1, problemNum = 1)

\S4method{getSigmas}{NMSimModel}(obj, what = "final", subProblemNum = 1,
  method = 1, problemNum = 1)

\S4method{getSigmas}{NMSimModelNM7}(obj, what = "final", subProblemNum = 1,
  method = 1, problemNum = 1)

\S4method{getSigmas}{NMSimDataGen}(obj, what = "final", subProblemNum = 1,
  method = 1, problemNum = 1)
}
\arguments{
\item{obj}{An object of class NMBasicModel, NMRun, NMSimModel, NMBasicModelNM7, NMSimModelNM7, or nmModel}

\item{what}{Character vector of items to extract. One or more of "final", "stderrors" or "initial" (or "shrinkage" for NONMEM 7 basic models)}

\item{subProblemNum}{Numeric vector of simulation sub-problems to use.  Only applies to simulation models}

\item{method}{Numeric vector of methods to extract when dealing with NONMEM 7 problems}

\item{problemNum}{Number of problem to reference - applies to runs only}
}
\value{
For NONMEM 6 - either a list of matrices if multiple "what=" are used, or a single matrix otherwise
For NONMEM 7 - The same as above for a single method.  If multiple methods, a list of lists or a list of matrices
}
\description{
A generic function that extracts sigma estimates (and initial estimates and standard errors if specified) from a NONMEM object.
Shrinkage values may also be retrieved for NONMEM 7 objects.
}
\note{
Invalid \code{what} elements are simply ignored.
}
\examples{
\dontrun{
     x <- importNm("theoph.con", path = "examples/theoph")
     getSigmas(x, what = c("initial", "final")) 
}
}
\author{
Mango Solutions
}
\keyword{methods}
