% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveMisc.R
\docType{methods}
\name{getSimInfo}
\alias{getSimInfo}
\alias{getSimInfo,NMRun-method}
\alias{getSimInfo,NMSimModelNM7-method}
\alias{getSimInfo,NMSimDataGen-method}
\alias{getSimInfo,NMSimModel-method}
\title{Get Simulation Information}
\usage{
getSimInfo(obj, problemNum = 1, addRawInfo = TRUE)

\S4method{getSimInfo}{NMRun}(obj, problemNum = 1, addRawInfo = TRUE)

\S4method{getSimInfo}{NMSimModelNM7}(obj, problemNum = 1, addRawInfo = TRUE)

\S4method{getSimInfo}{NMSimDataGen}(obj, problemNum = 1, addRawInfo = TRUE)

\S4method{getSimInfo}{NMSimModel}(obj, problemNum = 1, addRawInfo = TRUE)
}
\arguments{
\item{obj}{Object of class NMRun, or NMSim*}

\item{problemNum}{Number of problem (applicable only when obj is of class NMRun)}

\item{addRawInfo}{Should the raw contents of the $SIM statement be added to the returned information?}
}
\value{
A numeric vector with 3 entries : numSimulations, seed1, and seed2.  These are self-explanatory.
If addRawInfo is TRUE, the returned result will have an attribute named "rawStatement" that will have the text
of the control stream $SIM field.
}
\description{
Extract information simulation information from a simulation problem
}
\author{
Mango Solutions
}
\keyword{classes,}
\keyword{manip,}
\keyword{utilities}
