\name{getVarDescription}
\alias{getVarDescription}
\title{Gets a variable description}
\description{Gets the label, type and format of a selection of user specified variables, e.g. SEX and IPRED.}
\usage{
getVarDescription(varName)
}
\arguments{
  \item{varName}{A vector of variable names}
}
\value{A data frame of required variable names, labels, types and formats}
\author{Mango Solutions <rweeks@mango-solutions.com>}
\keyword{environment}