% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNmData.R
\name{readNmData}
\alias{readNmData}
\title{Read NONMEM dataset removing any "IGNORE" rows and dealing with (possibly repeated) headers 
This function is designed to import either output table or raw input data
The ignore string can optionally be added to remove rows from the data
Any table file headers found are removed from the data - repeated headers are also removed}
\usage{
readNmData(file, ignore = NULL, accept = NULL, translate = NULL,
  records = NULL, null = NA, sep = NULL)
}
\arguments{
\item{file}{Name of the file}

\item{ignore}{[C,1] - ignore option, see NM?$DATA[IGNORE]}

\item{accept}{[C,1] - accept option, see NM?$DATA[ACCEPT]}

\item{translate}{[C,1] - translate option, see NM?$DATA[TRANSLATE]}

\item{records}{[CN,1] - number of records, see NM?$DATA[RECORDS]}

\item{null}{[C,1] - value to replace NA, see NM?$DATA[NULL]}

\item{sep}{character that separates fields}
}
\value{
A data.frame of the contents of the tables in the file
}
\description{
Read NONMEM dataset removing any "IGNORE" rows and dealing with (possibly repeated) headers 
This function is designed to import either output table or raw input data
The ignore string can optionally be added to remove rows from the data
Any table file headers found are removed from the data - repeated headers are also removed
}
\examples{
 \dontrun{ 
     inData <- readNmData(file =  "examples/theoph/data.csv") 
     print(head(inData))
     outData <- readNmData(file = "examples/theoph/Outtable" )
     print(head(outData))
 }
}
\author{
Mango Solutions
}
\keyword{IO}
