\name{removeNmPath}
\alias{removeNmPath}
\title{Removes a user specified path}
\description{Removes a user specified path which has been stored globally under a path name.}
\usage{
removeNmPath(pathName)
}
\arguments{
  \item{pathName}{The path name}
}
\value{None}
\author{Mango Solutions <rweeks@mango-solutions.com>}
\keyword{environment}
