% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRNMImportTests.R
\name{runRNMImportTests}
\alias{runRNMImportTests}
\title{Run unit tests.}
\usage{
runRNMImportTests(TestPath = system.file(package = "RNMImport", "unittests"),
  ExcludeFolders = NULL, TestResult = NULL, ResultsType = c("html",
  "text"))
}
\arguments{
\item{TestPath}{Path of the folder which contains the test scripts.}

\item{ExcludeFolders}{The folders are not tested.}

\item{TestResult}{Name of the report file.}

\item{ResultsType}{'html' or 'text'.}
}
\value{
The results of function \code{\link{[RUnit]runTestSuite}}.
}
\description{
Run unit tests.
}
\details{
Run the unit tests by RUnit package, and generate a html or text report.
}
\examples{
\dontrun{
x <- runRNMImportTests(TestResult = "runRNMImportTests_tests")
summary(x)
}

}
\author{
Mango Solutions
}
\keyword{debugging}

