\name{setNmPath}
\alias{setNmPath}
\title{Stores a user specified path}
\description{Stores a path globally as a user specified path name.}
\usage{
setNmPath(pathName, path)
}
\arguments{
  \item{pathName}{The path name}
  \item{path}{The path}
}
\value{None}
\author{Mango Solutions <rweeks@mango-solutions.com>}
\keyword{environment}
