% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_methods.R
\docType{methods}
\name{showOutput}
\alias{showOutput}
\alias{show,NMRun-method}
\alias{show,NMBasicModel-method}
\alias{show,NMSimDataGen-method}
\alias{show,NMSimModel-method}
\alias{show,NMSimModelNM7-method}
\title{Show output table info}
\usage{
showOutput(object)

\S4method{show}{NMRun}(object)

\S4method{show}{NMBasicModel}(object)

\S4method{show}{NMSimDataGen}(object)

\S4method{show}{NMSimModel}(object)

\S4method{show}{NMSimModelNM7}(object)
}
\arguments{
\item{object}{NMProblem object}
}
\value{
none
}
\description{
Displays information about output data.  Utility function for use with show methods
}
\author{
Mango Solutions
}
